services/rest_api/requirements.txt
--pre
connexion[flask,swagger-ui,uvicorn]
pyjwt
peewee
mysqlclient
